src/requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0.1
WeasyPrint
Markdown
Jinja2
PyPDF2
reportlab
instagrapi
moviepy
trimesh
colorlog
emoji
pydal>=20.5